FLIM post processing tool
This is a GUI based tool for post processing of
SPCImage(http://www.becker-hickl.com/) FLIM data. As FLIM analysis poses
several challenges in terms of post processing, this tool aims to
minimize the effort. User can optimiza image based on chi^2 and
intensity value, calculate mean based on the valid pixel and export to
excel. This is a tool I use for my data. But you have to use a licensed
SPCImage to export data first.